commonly used functions moved to utils.py
